requirements.txt
streamlit
crewai
langchain
langgraph
chromadb
openai
python-dotenv
